Correccion de Vehiculos, IdentificadorVehiculos

+++ originacion/src/main/java/com/banquito/sistema/originacion/controller/IdentificadorVehiculoController.java
@@ -24,7 +24,7 @@ public class IdentificadorVehiculoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<IdentificadorVehiculo> getById(@PathVariable Integer id) {
+    public ResponseEntity<IdentificadorVehiculo> getById(@PathVariable Long id) {
         return ResponseEntity.ok(service.getById(id));
     }
