+++ originacion/src/main/java/com/banquito/sistema/originacion/exception/VehiculoNotFoundException.java
@@ -3,7 +3,7 @@ package com.banquito.sistema.originacion.exception;
 public class VehiculoNotFoundException extends RuntimeException {
     private final Integer errorCode;
 
-    public VehiculoNotFoundException(Integer id) {
+    public VehiculoNotFoundException(Long id) {
         super("Vehiculo not found with id " + id);
         this.errorCode = 1001;
     }
